Extract BatteryStatsDbDataSource.kt from BatteryStatsRepositoryImpl.kt
