--- batterysync/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/di/FeatureDI.kt
@@ -1,22 +1,0 @@
-package com.boswelja.smartwatchextensions.batterysync
-
-import com.boswelja.smartwatchextensions.batterysync.data.repository.BatteryStatsDbRepository
-import com.boswelja.smartwatchextensions.batterysync.domain.repository.BatteryStatsRepository
-import com.boswelja.smartwatchextensions.batterysync.database.BatteryStatsDatabase
-import org.koin.core.parameter.parametersOf
-import org.koin.core.qualifier.named
-import org.koin.dsl.module
-
-/**
- * A Koin module for providing Battery Sync repositories.
- */
-val batterySyncCommonModule = module {
-    single {
-        BatteryStatsDatabase(
-            get {
-                parametersOf(BatteryStatsDatabase.Schema, "batterystats.db")
-            }
-        )
-    }
-    single<BatteryStatsRepository> { BatteryStatsDbRepository(get(), get(named("database"))) }
-}
